major update

+++ src/gestionecampionati/grafica/Listener/ApriLogoActionListener.java
@@ -5,33 +5,29 @@
  */
 package gestionecampionati.grafica.Listener;
 
-import gestionecampionati.Squadra;
 import gestionecampionati.grafica.MainFrame;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import static java.lang.System.out;
 import java.nio.channels.FileChannel;
-import java.nio.file.Files;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import javax.imageio.ImageIO;
 import javax.swing.JFileChooser;
+import javax.swing.JFrame;
 
 /**
  *
  * @author enrico
  */
 public class ApriLogoActionListener implements ActionListener{
-    private MainFrame fr;
+    private JFrame fr;
     private String path;
 
-    public ApriLogoActionListener(MainFrame f) {
+    public ApriLogoActionListener(JFrame f) {
         this.fr = fr;
         this.path = new String("");
         
@@ -56,27 +52,27 @@ public class ApriLogoActionListener implements ActionListener{
                 FileChannel dest = null;
 
                 File f2 = new File("../GestioneCampionati/media/loghi/"+ f.getName());
+                if(!(f2.exists())){
+                        try {
+                            source = new FileInputStream(f).getChannel();
+                            dest = new FileOutputStream(f2).getChannel();
 
-                try {
-                    source = new FileInputStream(f).getChannel();
-                    dest = new FileOutputStream(f2).getChannel();
-
-                    if(source != null && dest != null){
-                        dest.transferFrom(source, 0, source.size());
-                        source.close();
-                        dest.close();
-
-
-
-                    }
+                            if(source != null && dest != null ){
+                                System.out.print("fefefefef");
+                                dest.transferFrom(source, 0, source.size());
+                                source.close();
+                                dest.close();
 
 
-                } catch (FileNotFoundException ex) {
-                    Logger.getLogger(ApriLogoActionListener.class.getName()).log(Level.SEVERE, null, ex);
-                } catch (IOException ex) {
-                    Logger.getLogger(ApriLogoActionListener.class.getName()).log(Level.SEVERE, null, ex);
+                            }
+
+
+                        } catch (FileNotFoundException ex) {
+                            Logger.getLogger(ApriLogoActionListener.class.getName()).log(Level.SEVERE, null, ex);
+                        } catch (IOException ex) {
+                            Logger.getLogger(ApriLogoActionListener.class.getName()).log(Level.SEVERE, null, ex);
+                        }
                 }
-        
         }catch(NullPointerException exc){
             
             System.out.println("Logo non inserito");
